refactor: chat/[id] 페이지 로직을 useChattingRoom 훅으로 분리
